Make the linenumbers config.toml changes opt-in
Unfortunately the config.toml is shared by all the examples and so we
try to push example-specific config into the index files. With markdown
configuation that's not possible, so the best we can do is tell the user
they need to comment it in to try it out. The better solution in the
future would be to pull this example out into it's own Hugo site.