adds mermaid support
Leverage the new codeblock renderhooks to support mermaid rendering.

Mermaid is picky about not being visible when it is rendered so we hook
on Reveal events to render on slide view or on ready event.

Mermaid is only included if it is used in the slides.

this change depends on hugo verison >= v0.93.0