adds initial render of content if it has mermaid
Before this change when the current page contained mermaid content
and was reloaded, we did not have any hooks to trigger the render trick.
(the events `slidechanged` or `ready` do not fire).

This commit adds an initial render call to ensure that mermaid elements
are processed if they are found on the current page.

Also are more in-line with other snippets by adding text/javascript
attribute.