adds empty default layouts for list and single
This hushes warnings about kind "page" (which is the default when no
kind is set).

The warning:

  WARN 2022/11/08 10:20:22 found no layout file for "HTML" for kind "page":
  You should create a template file which matches Hugo Layouts Lookup Rules
  for this combination.

Adding an empty default layout for list and single fixes this.

The single or list layout could have been created with a default html
comment:

{{ printf "<!-- Layout with no content to avoid WARN message about missing
page layout. Source: %s -->" . | safeHTML }}

This specifies the source that would have created the warning
but this causes outputing non-presentation related endpoints like
/categories or /home which has no relevance in a reveal-hugo context.
Even though these files are harmless.

This commit opted for the empty file variant in order to not pullute the
built site.